+++ kms-key-replica/variables.tf
@@ -1,0 +1,15 @@
+variable "project" {
+  type = string
+}
+
+variable "alias" {
+  type = string
+}
+
+variable "environment" {
+  type = string
+}
+
+variable "master_kms_key_arn" {
+  type = string
+}