+++ rds/outputs.tf
@@ -2,3 +2,7 @@
 output "database_url" {
   value = aws_db_instance.main.endpoint
 }
+
+output "database_arn" {
+  value = aws_db_instance.main.arn
+}