+++ rds/role-enhanced-monitoring.tf
@@ -1,6 +1,6 @@
 # RDS Enhanced Monitoring requires a specific role + KMS key to operate
 resource "aws_iam_role" "rds-enhanced-monitoring" {
-  name               = "rds-enhanced-monitoring-${var.project}-${var.environment}"
+  name               = "rds-enhanced-monitoring-${local.name}"
   assume_role_policy = data.aws_iam_policy_document.rds_enhanced_monitoring.json
 
   tags = {