+++ stack/app/outputs.tf
@@ -2,6 +2,10 @@ output "database_url" {
   value = module.rds.database_url
 }
 
+output "database_arn" {
+  value = module.rds.database_arn
+}
+
 output "redis_url" {
   value = module.elasticache.endpoint
 }
@@ -28,4 +32,18 @@ output "alb_target_group_ecs_arn" {
 }
 output "nlb_target_group_ecs_arn" {
   value = module.ecs.nlb_target_group_ecs_arn
+}
+
+
+# When launching a stack with a read replica
+output "accept_status-requester" {
+  value = var.rds_is_read_replica ? join("", module.vpc-peering.*.accept_status-requester) : "VPC peering not enabled."
+}
+
+output "accept_status-accepter" {
+  value = var.rds_is_read_replica ? join("", module.vpc-peering.*.accept_status-accepter) : "VPC peering not enabled."
+}
+
+output "rds_kms_key_arn" {
+  value = var.rds_master_db_kms_key_arn == null ? join("", module.rds-kms-key.*.arn) : "KMS ARN only printed for the master DB to be passed to each replica."
 }